requirements.txt
fastapi==0.115.5
pydantic==2.10.3
uvicorn==0.32.1
python-multipart==0.0.20
requests==2.31.0
Pillow==10.1.0
google-cloud-vision==3.7.0
google-generativeai==0.3.2
python-dotenv==1.0.0
